+++ packages/cooking/util/is.js
@@ -11,3 +11,4 @@ exports.Array = obj => type(obj) === '[object Array]'
 exports.Object = obj => type(obj) === '[object Object]'
 exports.Boolean = obj => type(obj) === '[object Boolean]'
 exports.Function = obj => type(obj) === '[object Function]'
+exports.nil = obj => obj === null || obj === undefined